+++ include/utils/binary.h
@@ -9,63 +9,68 @@
 #define _UTILS_BINARY_H_
 
 #include <fstream>
+#include <string>
 
-class BinaryFile {
+class BinaryReader {
   public:
-    BinaryFile(const char* f = nullptr);
-    ~BinaryFile();
+    virtual ~BinaryReader();
 
-    int open(const char* f = nullptr);
+    virtual long long tell() = 0;
+    virtual void seek(long long pos = 0) = 0;
+    virtual bool eof() = 0;
 
-    long long tell();
-    void seek(long long pos = 0);
+    virtual int8_t read8();
+    virtual uint8_t readU8();
 
-    int8_t read8();
-    uint8_t readU8();
+    virtual int16_t read16();
+    virtual uint16_t readU16();
 
-    int16_t read16();
-    uint16_t readU16();
+    virtual int32_t read32();
+    virtual uint32_t readU32();
 
-    int32_t read32();
-    uint32_t readU32();
+    virtual int64_t read64();
+    virtual uint64_t readU64();
 
-    int64_t read64();
-    uint64_t readU64();
-
-    float readFloat();
+    virtual float readFloat();
 
   private:
+    virtual void read(char* d, int c) = 0;
+};
+
+class BinaryFile : public BinaryReader {
+  public:
+    BinaryFile(std::string f = "");
+    virtual ~BinaryFile();
+
+    int open(std::string f = "");
+
+    virtual long long tell();
+    virtual void seek(long long pos = 0);
+    virtual bool eof();
+
+  private:
+    virtual void read(char* d, int c);
+
     std::ifstream file;
 };
 
-class BinaryMemory {
+class BinaryMemory : public BinaryReader {
   public:
-    BinaryMemory(char* d = nullptr);
+    BinaryMemory(const char* d = nullptr, long long max = -1);
+    virtual ~BinaryMemory();
 
-    int open(char* d = nullptr);
+    int open(const char* d = nullptr, long long max = -1);
 
-    long long tell();
-    void seek(long long pos = 0);
-
-    int8_t read8();
-    uint8_t readU8();
-
-    int16_t read16();
-    uint16_t readU16();
-
-    int32_t read32();
-    uint32_t readU32();
-
-    int64_t read64();
-    uint64_t readU64();
-
-    float readFloat();
+    virtual long long tell();
+    virtual void seek(long long pos = 0);
+    virtual bool eof();
 
   private:
-    void read(char* d, int c);
+    virtual void read(char* d, int c);
 
-    char* data;
+    const char* data;
     long long offset;
+    long long max;
 };
 
 #endif